refactor(python-sdk): move Params from llm.providers.base to llm.models
It's been bugging me that params are a core type to model but the
implementation is buried alongside the base provider.

While here:
- Split thinking_config.py into its own file
- Expose more symbols in llm so that ops imports can be cleaner
