+++ python/mirascope/llm/providers/anthropic/provider.py
@@ -1,6 +1,9 @@
 """Anthropic client implementation."""
 
+from __future__ import annotations
+
 from collections.abc import Sequence
+from typing import TYPE_CHECKING
 from typing_extensions import Unpack
 
 from anthropic import Anthropic, AsyncAnthropic
@@ -28,11 +31,14 @@ from ...tools import (
     Tool,
     Toolkit,
 )
-from ..base import BaseProvider, Params
+from ..base import BaseProvider
 from . import _utils
 from .beta_provider import AnthropicBetaProvider
 from .model_id import AnthropicModelId, model_name
 from .model_info import MODELS_WITHOUT_STRICT_STRUCTURED_OUTPUTS
+
+if TYPE_CHECKING:
+    from ...models import Params
 
 
 def _should_use_beta(
